Revert "feat: validation for dx address is modified and code committed. (#592)"
This reverts commit 3fc33b87257cdf4dd4b49d40c21d3107fe8c34c6.
