+++ src/integrationTest/java/uk/gov/hmcts/reform/professionalapi/EditPaymentAccountsTest.java
@@ -90,7 +90,7 @@ public class EditPaymentAccountsTest extends AuthorizationEnabledIntegrationTest
         java.util.Map<String, Object> retrievePaymentAccountsByEmailResponse = professionalReferenceDataClient.retrieveSingleOrganisation(orgId, hmctsAdmin);
 
         List paymentAccounts = (List) retrievePaymentAccountsByEmailResponse.get("paymentAccount");
-        assertThat(paymentAccounts).isNull();
+        assertThat(paymentAccounts).hasSize(0);
     }
 
     @Test
