+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/response/ContactInformationResponse.java
@@ -27,7 +27,7 @@ public class ContactInformationResponse {
     @JsonProperty
     private final String postCode;
     @JsonProperty
-    private  List<DxAddressResponse> dxAddress;
+    private final List<DxAddressResponse> dxAddress;
 
     public ContactInformationResponse(ContactInformation contactInfo) {
         this.addressLine1 = contactInfo.getAddressLine1();
@@ -37,12 +37,10 @@ public class ContactInformationResponse {
         this.county = contactInfo.getCounty();
         this.country = contactInfo.getCountry();
         this.postCode = contactInfo.getPostCode();
-        if (contactInfo.getDxAddresses().size() > 0) {
-            this.dxAddress = contactInfo.getDxAddresses()
-                    .stream()
-                    .map(dxAddres -> new DxAddressResponse(dxAddres))
-                    .collect(toList());
-        }
+        this.dxAddress = contactInfo.getDxAddresses()
+                .stream()
+                .map(dxAddres ->  new DxAddressResponse(dxAddres))
+                .collect(toList());
     }
 
 }
