+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/response/OrganisationEntityResponse.java
@@ -49,12 +49,10 @@ public class OrganisationEntityResponse  {
         if (!organisation.getUsers().isEmpty()) {
             this.superUser = new SuperUserResponse(organisation.getUsers().get(0));
         }
-        if (organisation.getPaymentAccounts().size() > 0) {
-            this.paymentAccount = organisation.getPaymentAccounts()
-                    .stream()
-                    .map(pbaAccount -> new PbaAccountResponse(pbaAccount).getPbaNumber())
-                    .collect(toList());
-        }
+        this.paymentAccount = organisation.getPaymentAccounts()
+                .stream()
+                .map(pbaAccount -> new PbaAccountResponse(pbaAccount).getPbaNumber())
+                .collect(toList());
         if (Boolean.TRUE.equals(isRequiredAllEntities)) {
             this.contactInformation = organisation.getContactInformation()
                     .stream()
